requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Streamlit for UI
streamlit==1.28.0

# AWS dependencies
boto3==1.34.0
botocore==1.34.0

# Machine Learning and NLP - CPU-only versions
# Install PyTorch CPU-only version to save space
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.1+cpu
transformers==4.35.0
tokenizers==0.14.1

# Vector search and similarity
faiss-cpu==1.7.4
numpy==1.24.3

# Data processing
openpyxl==3.1.2
pandas==2.1.4
scikit-learn==1.3.2

# Text processing and fuzzy matching
rapidfuzz==3.5.2
nltk==3.8.1

# HTTP client
httpx==0.25.2
requests==2.31.0

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1

# Logging and monitoring
structlog==23.2.0

# Process management
supervisor==4.2.5

# Development and testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Additional utilities
Pillow==10.1.0
python-json-logger==2.0.7
colorama==0.4.6
